Udbs 20171118 threeboy

+++ ft.serverone/src/main/java/com/future/technology/serverone/essay/service/IEssayService.java
@@ -2,10 +2,8 @@ package com.future.technology.serverone.essay.service;
 
 import com.future.technology.serverone.common.Response;
 import com.future.technology.serverone.essay.domain.Essay;
-import com.future.technology.serverone.essay.domain.PageBean;
 import com.future.technology.serverone.essay.domain.QueryInfo;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import com.future.technology.serverone.utils.PageBean;
 
 import java.util.List;
 
@@ -22,7 +20,7 @@ public interface IEssayService {
 
     Response updateEssay(Essay essay);
 
-    Response downline(Integer essay_id);
+    Response downline(Long essay_id);
 
     Response<PageBean> queryEssay(QueryInfo queryInfo);
